requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.5.1
bandit==1.8.3
bcrypt==4.2.1
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
dnspython==2.6.1
dparse==0.6.4
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
identify==2.6.9
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
nltk==3.9.1
nodeenv==1.9.1
packaging==24.2
pbr==6.1.1
platformdirs==4.3.7
py-machineid==0.7.0
pre_commit==4.2.0
psutil==6.1.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.8.0
ruamel.yaml==0.18.10
safety==3.3.1
safety-schemas==0.0.11
setuptools==78.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
stevedore==5.4.1
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.30.6
virtualenv==20.29.3
watchfiles==0.24.0
websockets==13.0.1
